backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
pymongo==4.8.0
dnspython==2.8.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
groq==0.11.0
python-multipart==0.0.12
slowapi==0.1.9
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
ruff==0.6.0
mypy==1.11.0
